Add scale functionality
Allows the loaded polygon to be rescaled. If you plan on using different sized Sprites, you would probably want to use different sized bodies. The scale can't be changed later, but that wouldn't matter if the Sprite using this polygon has a constant scale. I've used this reworked method for a while and it has been working perfectly so far, even with negative numbers.