Add powerPreference hint to WebGL context options
This adds powerPreference hinting to the PIXI WebGL renderer's context
options, per the spec at
https://www.khronos.org/registry/webgl/specs/latest/1.0/. In situations
where a high-performance GPU is available (e.g. laptops with discrete
GPUs) this offers the ability to hint to the browser that you would
like to use one GPU or the other. Note that this may cause greater
power consumption.

Per the WebGL spec, "Applications that request high-performance should
test and maintain robust context loss handling, as User Agents are very
likely to decide to lose background high-performance contexts." Phaser
appears to handle context loss.

Phaser 3.x has this setting in game config (/src/core/Config.js), and
WebGLRenderer.js L581 instantiates the context in a similar fashion to
this pull request. This is an attempt to backport the function to
the Phaser CE rendering context.