TilemapLayer: fix getRayCastTile implementation
Previously, getRayCastTile relied on getTiles, which retrieves
a list of tiles in some rectangle. Not only was this inefficient,
but it would cause getRayCastTile to sometimes return an empty
array when there were tiles in the ray. Specifically, this bug
could be triggered by calling getRayCastTile on a horizontal
or vertical line. Since this is converted to a zero width or height,
getTiles would return an empty array because th = 0 or tw = 0 in the
for statements:

for (var wy = ty; wy < ty + th; wy++)

and

for (var wx = tx; wx < tx + tw; wx++)

meaning these loops would not even start. This behavior makes sense
for the getTiles function, but does not for getRayCastTiles. Instead,
we get an array of coordinates on the line to check, and simply check
if a tile is present at these coordinates.